+++ docs/update-configuration.js
@@ -119,6 +119,7 @@ const props = [
           break;
         case 'function': {
           let fixIndent;
+          let mute = false;
           append(String(value).split('\n').map((line, index) => {
             if (index === 1) {
               line.match(/^(\s+)\S+/);
@@ -128,6 +129,15 @@ const props = [
             if (line.startsWith(' ')) line = line.slice(fixIndent);
             line = line.replace(/ \/\/ eslint-disable.+/, '');
             line = line.replace(/ \/\/ TODO.+/, '');
+            line = line.replace(/ class="[ \-\w]+ ?"/, '');
+            if (line.includes('<style>')) {
+              mute = true;
+              return '<style>/* css and html classes omitted for brevity, see lib/helpers/defaults.js */</style>';
+            } else if (line.includes('</style>')) {
+              mute = false;
+              return undefined;
+            }
+            if (mute) return undefined;
             return line;
           }).filter(Boolean)
             .join('\n'));