+++ example/settings.js
@@ -31,6 +31,7 @@ module.exports.config = {
 
     backchannelLogout: true, // defaults to false
     claimsParameter: true, // defaults to false
+    conformIdTokenClaims: true, // defaults to false
     encryption: true, // defaults to false
     frontchannelLogout: true, // defaults to false
     introspection: true, // defaults to false